requirements.txt
python-telegram-bot==22.1
supabase==2.15.2
bittensor==9.7.0
httpx==0.28.1
gotrue==2.12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
aiohttp==3.12.9
async-substrate-interface==1.2.2
pydantic==2.11.5
PyYAML==6.0.2
requests==2.32.3 